feat: 102-fibonacci.c task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,16 +8,16 @@
 int main(void)
 {
 int n = 50; /* Number of Fibonacci numbers to print*/
-long long int fib1 = 1, fib2 = 2, next;
+long int fib1 = 1, fib2 = 2, next;
 
 /* Print the first two Fibonacci numbers*/
-printf("%lld, %lld", fib1, fib2);
+printf("%ld, %ld", fib1, fib2);
 
 /* Calculate and print the remaining Fibonacci numbers*/
 for (int i = 3; i <= n; i++)
 {
 next = fib1 + fib2;
-printf(", %lld", next);
+printf(", %ld", next);
 
 /* Update values for the next iteration*/
 fib1 = fib2;
